Do not delete input if there is no tarball

+++ outsource/workflow/combine-wrapper.sh
@@ -122,8 +122,10 @@ echo
 echo "Total amount of data in $input_path before removing: "`du -s --si $input_path | cut -f1`
 echo
 
-echo "Removing inputs directory:"
-rm -r $input_path
+if [ $tar_filename != "X" ]; then
+    echo "Removing inputs directory:"
+    rm -r $input_path
+fi
 
 echo "Here is what is in the output directory after combining:"
 ls -lah $output_path
